Let ProjectType handle the xcodeproj instead
It seems like XCProjectFile expects a xcodeproj url. Before this patch a
pbxproject was being passed instead which cause XCProjectFile to fail
initialization.

Retrieving this is simple when a project is given to the daemon. I've
deprecated usage for workspace and folder options for now because I
don't see an easy way to handle them.