Use Carthage to build frameworks
Checkouts are being versioned because there is a slight error with
SwiftSockets where Carthage tries to build a non-shared scheme. This has
been manually fixed in the checkout.

Sourcekitten needs to be compiled as opposed to having Carthage
automatically download the binary because the binary comes without the
header files. See jpsim/SourceKitten#107