First draft of the FSEvents wrapper
Since the File System events APIs available for each OS are different, we're using conditional compilation blocks:
https://developer.apple.com/library/content/documentation/Swift/Conceptual/Swift_Programming_Language/Statements.html#//apple_ref/doc/uid/TP40014097-CH33-ID538

This compiles in Linux. If you're using the latest linux-sourcekit-support commits, you may have to fix the `as` expressions on XCode.swift:Sources/XCProjectFile.swift.
Those expressions throw a compiler error, but are fixable by just replacing all the ' as ' in the file for ' as! '.

Should compile (up to some minor errors if I misplaced some bracket) in Mac OS.

Untested:
- Actual functionality on Linux (that's the next step)
- Compilation and correct functionality on Mac OS (to fix any mistakes that I could've made there)